Update Action configurator screen to preview the same as the real widget
