+++ Sources/App/Settings/ActionConfigurator.swift
@@ -4,6 +4,7 @@ import Foundation
 import PromiseKit
 import RealmSwift
 import Shared
+import SwiftUI
 import UIKit
 import ViewRow
 
@@ -87,7 +88,7 @@ class ActionConfigurator: HAFormViewController, TypedRowControllerType {
         let visuals = Section()
 
         if action.canConfigure(\Action.Text) || action.isServerControlled {
-            let section: Section
+            let section: Eureka.Section
 
             if action.canConfigure(\Action.Text) {
                 section = visuals
@@ -300,65 +301,63 @@ class ActionConfigurator: HAFormViewController, TypedRowControllerType {
 }
 
 class ActionPreview: UIView {
-    var imageView = UIImageView(frame: CGRect(x: 15, y: 0, width: 44, height: 44))
-    var title = UILabel(frame: CGRect(x: 60, y: 60, width: 200, height: 100))
     var action: Action?
+
+    private let stackView = UIStackView()
 
     override func layoutSubviews() {
         super.layoutSubviews()
 
-        layer.cornerRadius = 8
-        let centerY = (frame.size.height / 2) - 50
+        addSubview(stackView)
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            stackView.topAnchor.constraint(equalTo: topAnchor),
+            stackView.bottomAnchor.constraint(equalTo: bottomAnchor),
+            stackView.leftAnchor.constraint(equalTo: leftAnchor),
+            stackView.rightAnchor.constraint(equalTo: rightAnchor),
+        ])
 
-        title = UILabel(frame: CGRect(x: 60, y: centerY, width: 200, height: 100))
-
-        title.textAlignment = .natural
-        title.clipsToBounds = true
-        title.numberOfLines = 1
-        title.font = title.font.withSize(UIFont.smallSystemFontSize)
-
-        addSubview(title)
-        addSubview(imageView)
-
-        let tap = UITapGestureRecognizer(target: self, action: #selector(handleGesture))
-        addGestureRecognizer(tap)
+        setup(action)
     }
 
-    public func setup(_ action: Action) {
+    func setup(_ action: Action?) {
         self.action = action
-        DispatchQueue.main.async {
-            self.backgroundColor = UIColor(hex: action.BackgroundColor)
 
-            let icon = MaterialDesignIcons(named: action.IconName)
-            self.imageView.image = icon.image(
-                ofSize: self.imageView.bounds.size,
-                color: UIColor(hex: action.IconColor)
-            )
-            self.title.text = action.Text
-            self.title.textColor = UIColor(hex: action.TextColor)
+        for subview in stackView.arrangedSubviews {
+            stackView.removeArrangedSubview(subview)
+            subview.removeFromSuperview()
         }
+        guard let action else { return }
+        let previewView = UIHostingController(rootView: WidgetPreviewView(action: action))
+        stackView.addArrangedSubview(previewView.view)
     }
+}
 
-    @objc func handleGesture(gesture: UITapGestureRecognizer) {
-        guard let action,
-              let server = Current.servers.server(forServerIdentifier: action.serverIdentifier) else {
-            return
+#Preview {
+    VStack {
+        WidgetPreviewView(action: Action())
+    }
+}
+
+struct WidgetPreviewView: View {
+    let action: Action
+    var body: some View {
+        VStack {
+            WidgetBasicView(
+                model: .init(
+                    id: action.ID,
+                    title: action.Text,
+                    subtitle: nil,
+                    interactionType: .widgetURL(URL(string: "https://google.com")!),
+                    icon: MaterialDesignIcons(named: action.IconName),
+                    backgroundColor: Color(uiColor: .init(hex: action.BackgroundColor))
+                ),
+                sizeStyle: .condensed
+            )
+            .padding()
+            .frame(width: 340, height: 100)
         }
-
-        let feedbackGenerator = UINotificationFeedbackGenerator()
-        feedbackGenerator.prepare()
-
-        imageView.showActivityIndicator()
-
-        firstly {
-            Current.api(for: server).HandleAction(actionID: action.ID, source: .Preview)
-        }.done { _ in
-            feedbackGenerator.notificationOccurred(.success)
-        }.ensure {
-            self.imageView.hideActivityIndicator()
-        }.catch { err in
-            Current.Log.error("Error during action event fire: \(err)")
-            feedbackGenerator.notificationOccurred(.error)
-        }
+        .frame(maxWidth: .infinity)
+        .background(Color(uiColor: .tertiarySystemGroupedBackground))
     }
 }
